add migration of room from version 2 to version 3
